+++ suite-native/storage/package.json
@@ -14,6 +14,7 @@
         "@mobily/ts-belt": "^3.13.1",
         "@reduxjs/toolkit": "2.8.2",
         "@sentry/react-native": "6.20.0",
+        "@suite-common/firmware-authenticity": "workspace:*",
         "@suite-common/suite-types": "workspace:*",
         "@suite-common/wallet-core": "workspace:*",
         "bs58": "^6.0.0",
