+++ suite-native/storage/tsconfig.json
@@ -3,6 +3,9 @@
     "compilerOptions": { "outDir": "libDev" },
     "references": [
         {
+            "path": "../../suite-common/firmware-authenticity"
+        },
+        {
             "path": "../../suite-common/suite-types"
         },
         {
